refactor: Split health-monitor into focused modules (#173)
Refactors the 1,013-line health-monitor.ts file into a modular structure
with clear separation of concerns. This is PR #4 of the refactoring plan
outlined in issue #173.

## Changes

### Structure
- **types.ts**: Type definitions (HealthCheckResult, HealthCheck, SystemMetrics, HealthCheckConfig)
- **core/**: Main health monitoring logic
  - health-monitor.ts: Simplified HealthMonitor class (~300 lines)
  - health-check-runner.ts: Health check orchestration
  - scoring.ts: Health scoring and recommendation logic
- **checks/**: Individual health check modules
  - memory-check.ts: Memory usage monitoring
  - logging-check.ts: Logging system health
  - request-check.ts: Request tracking health
  - performance-check.ts: Performance monitoring health
  - configuration-check.ts: Configuration system health
- **alerts/**: Alert management
  - alert-manager.ts: Alert sending and cooldown logic
- **middleware/**: HTTP server integration
  - http-middleware.ts: Express/HTTP health check endpoints
- **instances.ts**: Global instances and healthChecks helper
- **index.ts**: Public API (types + healthChecks)

### Breaking Changes
- Import paths updated: @/utils/logging re-exports remain the same
- Removed unused exports: HealthMonitor, globalHealthMonitor,
  healthCheckMiddleware, initializeHealthMonitoring (not used in codebase)
- Public API now only exports types and healthChecks utility

### Testing
- ✅ All 13 new modules have test coverage (25+ new tests)
- ✅ All 2545 existing tests pass
- ✅ Type checking passes
- ✅ No unused code (knip passes)
- ✅ Linting passes
