+++ source/utils/logging/health-monitor/checks/configuration-check.ts
@@ -1,0 +1,43 @@
+/**
+ * Configuration system health check
+ */
+
+import type {HealthCheck, HealthCheckConfig} from '../types.js';
+
+/**
+ * Calculate score based on status
+ */
+function calculateScore(status: 'pass' | 'fail' | 'warn'): number {
+	if (status === 'fail') return 0;
+	if (status === 'pass') return 100;
+	if (status === 'warn') return 50;
+	return 50; // default for unknown status
+}
+
+/**
+ * Check configuration system health
+ */
+export function checkConfigurationSystem(
+	_config: HealthCheckConfig,
+): HealthCheck {
+	const startTime = performance.now();
+	const reloaderStats = {isEnabled: false, watcherCount: 0};
+
+	const status: 'pass' | 'fail' | 'warn' = 'pass';
+	const message = 'Configuration system is healthy';
+	const details = {
+		isEnabled: reloaderStats.isEnabled,
+		watcherCount: reloaderStats.watcherCount,
+		listenerCount: 0,
+		hasPendingReloads: false,
+	};
+
+	return {
+		name: 'configuration-system',
+		status,
+		score: calculateScore(status),
+		duration: performance.now() - startTime,
+		message,
+		details,
+	};
+}
