+++ source/utils/logging/health-monitor/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Health monitoring system
+ * Provides comprehensive health metrics and monitoring capabilities
+ */
+
+// Re-export types
+export type {
+	HealthCheck,
+	HealthCheckConfig,
+	HealthCheckResult,
+	SystemMetrics,
+} from './types.js';
+
+// Re-export the only public API - healthChecks helper functions
+export {healthChecks} from './instances.js';
